+++ tests/test_cases/moved_groups.d
@@ -1,0 +1,54 @@
+build/rom.elf: \
+    build/src/main_segment/main.o \
+    build/src/main_segment/nnsched.o \
+    build/src/main_segment/joy.o \
+    build/src/libmus/player.o \
+    build/src/libmus/player_fx.o \
+    build/src/libmus/aud_dma.o \
+    build/src/libmus/aud_sched.o \
+    build/src/libmus/aud_thread.o \
+    build/src/libmus/lib_memory.o \
+    build/src/libmus/aud_samples.o \
+    build/src/stufsh/thingy1.o \
+    build/src/stufsh/thingy2.o \
+    build/src/stufsh/thonga3.o \
+    build/src/stufsh/capy4.o \
+    build/src/stufsh/placeholder5.o \
+    build/src/stufsh/idk6.o \
+    build/src/stufsh/another7.o
+
+build/src/main_segment/main.o:
+build/src/main_segment/nnsched.o:
+build/src/main_segment/joy.o:
+build/src/libmus/player.o:
+build/src/libmus/player_fx.o:
+build/src/libmus/aud_dma.o:
+build/src/libmus/aud_sched.o:
+build/src/libmus/aud_thread.o:
+build/src/libmus/lib_memory.o:
+build/src/libmus/aud_samples.o:
+build/src/stufsh/thingy1.o:
+build/src/stufsh/thingy2.o:
+build/src/stufsh/thonga3.o:
+build/src/stufsh/capy4.o:
+build/src/stufsh/placeholder5.o:
+build/src/stufsh/idk6.o:
+build/src/stufsh/another7.o:
+
+-include build/src/main_segment/main.d
+-include build/src/main_segment/nnsched.d
+-include build/src/main_segment/joy.d
+-include build/src/libmus/player.d
+-include build/src/libmus/player_fx.d
+-include build/src/libmus/aud_dma.d
+-include build/src/libmus/aud_sched.d
+-include build/src/libmus/aud_thread.d
+-include build/src/libmus/lib_memory.d
+-include build/src/libmus/aud_samples.d
+-include build/src/stufsh/thingy1.d
+-include build/src/stufsh/thingy2.d
+-include build/src/stufsh/thonga3.d
+-include build/src/stufsh/capy4.d
+-include build/src/stufsh/placeholder5.d
+-include build/src/stufsh/idk6.d
+-include build/src/stufsh/another7.d